+++ client-web/src/App.js
@@ -1,0 +1,9 @@
+import React from "react"
+
+function App() {
+    return (
+       <h1>test</h1>
+    );
+}
+
+export default App;
